% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acealanme}
\alias{acealanme}
\title{Hills from 30 ns metadynamics of AceAlaNme in water with two collective variables}
\format{hillsfile object}
\source{
\url{http://www.metadynamics.cz/metadynminer/data/HILLS2d}
}
\usage{
acealanme
}
\description{
Hills from 30 ns metadynamics of AceAlaNme (Amber99SB-ILDN) in water (TIP3P)
with Ramachandran dihedrals phi and psi as collective variables.
}
\keyword{datasets}
