\docType{data}
\name{acealanme1.3}
\alias{acealanme1.3}
\title{Hills data from AceAlaNme metadynamics in water}
\format{The object of class \code{hillsfile} with 7,000 hills sampled every 0.5 ps with widths 0.3 rad and height 0.5 kJ/mol}
\source{
cmputed by Plumed 1.3
}
\usage{
acealanme1.3
}
\description{
Hills data from AceAlaNme metadynamics in water in the space of Ramachandran angles computed by Plumed 1.3.
}
\examples{
\dontrun{
data(acealanme1.3)
summary(acealanme1.3)
plot(acealanme1.3)
}
}
\author{
Vojtech Spiwok
}
\keyword{datasets}
