% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acealanme3d}
\alias{acealanme3d}
\title{Hills from 30 ns metadynamics of AceAlaNme in water with three collective variable (phi, psi, omega)}
\format{hillsfile object}
\source{
\url{http://www.metadynamics.cz/metadynminer/data/HILLS3d}
}
\usage{
acealanme3d
}
\description{
Hills from 30 ns metadynamics of AceAlaNme (Amber99SB-ILDN) in water (TIP3P)
with a Ramachandran dihedral phi and psi and peptide bond omega as collective variables.
}
\keyword{datasets}
