% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{emptyhills}
\alias{emptyhills}
\title{Generate empty HILLS from Plumed}
\usage{
emptyhills(dim=2, per=c(FALSE, FALSE), pcv1=c(-pi,pi), pcv2=c(-pi,pi))
}
\arguments{
\item{dim}{dimensionality of the output.}

\item{per}{logical vector specifying periodicity of collective variables.}

\item{pcv1}{periodicity of CV1.}

\item{pcv2}{periodicity of CV2.}
}
\value{
hillsfile object.
}
\description{
`emptyhills` returns a hillsfile object with no hills.
User can specify whether some collective variables are periodic.
}
\examples{
emptyhills(dim=2)
}
