% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{head.hillsfile}
\alias{head.hillsfile}
\title{Print first n lines of hillsfile}
\usage{
\method{head}{hillsfile}(x, n = 10, ...)
}
\arguments{
\item{x}{hillsfile object.}

\item{n}{number of lines (default 10).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`head.hillsfile` prints first n lines of a hillsfile object.
}
\examples{
head(acealanme)
}
