% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{lines.fes}
\alias{lines.fes}
\title{Plots 1D free energy surface object as lines}
\usage{
\method{lines}{fes}(x, lwd = 1, col = "black", ...)
}
\arguments{
\item{x}{fes object.}

\item{lwd}{line width for drawing symbols see 'par'.}

\item{col}{color code or name, see 'par'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`lines.fes` plots 1D free energy surface as lines.
}
\examples{
tfes<-fes(acealanme1d, imax=5000)
plot(tfes)
lines(tfes, lwd=4)
}
