% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{lines.hillsfile}
\alias{lines.hillsfile}
\title{Plot lines for hillsfile object}
\usage{
\method{lines}{hillsfile}(hills = hills, ignoretime = FALSE, lwd = 1,
  col = "black", ...)
}
\arguments{
\item{hills}{hillsfile object}

\item{ignoretime}{time in the first column of the HILLS file will be ignored}
}
\description{
`lines.hillsfile` plots lines for hillsfile object. For a hillsfile with one
collective variable it plots its evolution. For a hillsfile with two collective
variables it plots CV1 vs. CV2
}
\examples{
plot(acealanme)
lines(acealanme, col="red")
}
