% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebs.R
\name{lines.nebpath}
\alias{lines.nebpath}
\title{Plot lines for Nudged Elastic Band}
\usage{
\method{lines}{nebpath}(nebpath = nebpath, col = "red", lwd = 1, ...)
}
\arguments{
\item{nebpath}{nebpath object}
}
\description{
`lines.nebpath` plots lines for free energy profile calculated by Nudged Elastic Band.
}
\examples{
tfes<-fes(acealanme)
minima<-fesminima(tfes)
nebAD<-neb(minima, min1="A", min2="D")
plot(nebAD)
lines(nebAD, lwd=4)
}
