% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{max.fes}
\alias{max.fes}
\title{Calculate maximum of free energy surface}
\usage{
\method{max}{fes}(inputfes, na.rm = NULL, ...)
}
\arguments{
\item{inputfes}{fes object.}

\item{na.rm}{a logical indicating whether missing values should be
removed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`max.fes` calculates maximum of free energy in a fes object.
}
\examples{
tfes<-fes(acealanme, imax=5000)
max(tfes)
}
