% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{mean.fes}
\alias{mean.fes}
\title{Calculate mean of free energy surface}
\usage{
\method{mean}{fes}(inputfes = inputfes, na.rm = NULL)
}
\arguments{
\item{inputfes}{fes object}
}
\value{
mean
}
\description{
`mean.fes` calculates mean of free energy in a fes object
}
\examples{
tfes<-fes(acealanme)
mean(tfes)
}
