\name{metadynminer-package}
\alias{metadynminer-package}
\alias{metadynminer}
\docType{package}
\title{
\packageTitle{metadynminer}
}
\description{
\packageDescription{metadynminer}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{metadynminer}
\packageIndices{metadynminer}
}
\author{
\packageAuthor{metadynminer}

Maintainer: \packageMaintainer{metadynminer}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
