% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{min.fes}
\alias{min.fes}
\title{Calculate minimum of free energy surface}
\usage{
\method{min}{fes}(inputfes, na.rm = NULL, ...)
}
\arguments{
\item{inputfes}{fes object.}

\item{na.rm}{a logical indicating whether missing values should be
removed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`min.fes` calculates minimum of free energy in a fes object.
}
\examples{
tfes<-fes(acealanme)
min(tfes)
}
