% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{oneminimum}
\alias{oneminimum}
\title{Creates one ad hoc free energy minimum for a fes object (generic function
for 'metadynminer' and 'metadynminer3d')}
\usage{
oneminimum(inputfes, cv1, cv2)
}
\arguments{
\item{inputfes}{fes object.}

\item{cv1}{the value of collective variable 1.}

\item{cv2}{the value of collective variable 2.}
}
\value{
minima object.
}
\description{
`oneminimum` creates an ad hoc free energy minimum on free energy surface.
This can be used to calculate free energy surface evolution at arbitrary
point of free energy surface.
}
