% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{plot.fes}
\alias{plot.fes}
\title{Plot fes object}
\usage{
\method{plot}{fes}(inputfes = inputfes, plottype = "both",
  colscale = F, x = NULL, y = NULL, xlim = NULL, ylim = NULL,
  zlim = NULL, main = NULL, sub = NULL, xlab = NULL, ylab = NULL,
  nlevels = 10, levels = NULL, col = rainbow(135)[100:1],
  labels = NULL, labcex = 0.6, drawlabels = TRUE,
  colscalelab = "free energy", method = "flattest",
  contcol = par("fg"), lty = par("lty"), lwd = 1, axes = T)
}
\arguments{
\item{inputfes}{fes object}

\item{plottype}{specifies whether 2D free energy surface will be ploted as image, contours or both (default "both")}

\item{colscale}{specifies whether color scale will be ploted (default False)}

\item{colscalelab}{color scale label (default "free energy")}
}
\description{
`plot.fes` plots free energy surface. For a fes with one collective variable it plots
a 1D profile. For a fes with two collective variables it plots 2D free energy surface
using image, contours or combination of both (default).
}
\examples{
tfes2d<-fes(acealanme)
plot(tfes2d)
tfes1d<-fes(acealanme1d)
plot(tfes1d)
}
