% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{plot.minima}
\alias{plot.minima}
\title{Plot minima object}
\usage{
\method{plot}{minima}(minims = minims, plottype = "both", x = NULL,
  y = NULL, xlim = NULL, ylim = NULL, zlim = NULL, main = NULL,
  sub = NULL, xlab = NULL, ylab = NULL, nlevels = 10,
  levels = NULL, col = rainbow(135)[100:1], labels = NULL,
  labcex = 0.6, drawlabels = TRUE, method = "flattest",
  textcol = "black", pch = 1, bg = "red", cex = 1,
  contcol = par("fg"), lty = par("lty"), lwd = par("lwd"),
  axes = T)
}
\arguments{
\item{minims}{minima object}

\item{plottype}{specifies whether 2D free energy surface will be ploted as image, contours or both (default "both")}

\item{colscale}{specifies whether color scale will be ploted (default False)}

\item{colscalelab}{color scale label (default "free energy")}
}
\description{
`plot.minima` plots free energy surface with minima. The free energy sufrace is ploted the same
way as by plot.fes with aditional minima labels.
}
\examples{
tfes<-fes(acealanme)
minima<-fesminima(tfes)
plot(minima)
}
