% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{plotheights}
\alias{plotheights}
\title{Plot evolution of heights of hills in hillsfile object}
\usage{
plotheights(hills = hills, ignoretime = FALSE, xlab = NULL,
  ylab = NULL, xlim = NULL, ylim = NULL, zlim = NULL,
  main = NULL, sub = NULL, pch = 1, col = "black", bg = "red",
  cex = 1, asp = NULL, lwd = 1, axes = TRUE)
}
\arguments{
\item{hills}{hillsfile object}

\item{ignoretime}{time in the first column of the HILLS file will be ignored}
}
\description{
`plotheights` plots evolution of heights of hills. In well tempered metadynamics
hill heights decrese with flooding of the free energy surface. Evolution of heights
may be useful to evaluate convergence of the simulation.
}
\examples{
plotheights(acealanme)
}
