% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{points.fes}
\alias{points.fes}
\title{Plots 1D fes object as points}
\usage{
\method{points}{fes}(inputfes = inputfes, x = NULL, pch = 1,
  col = "black", bg = "red", cex = 1)
}
\arguments{
\item{inputfes}{fes object}
}
\description{
`points.fes` plots 1D free energy surface as points.
}
\examples{
tfes<-fes(acealanme1d)
plot(tfes)
points(tfes)
}
