% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{print.fes}
\alias{print.fes}
\title{Print dimensionality, minimum and maximum of free energy surface}
\usage{
\method{print}{fes}(inputfes = inputfes)
}
\arguments{
\item{inputfes}{fes object}
}
\description{
`print.fes` prints dimensionality, minimum and maximum of
free energy in a fes object
}
\examples{
tfes<-fes(acealanme)
tfes
}
