% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{print.hillsfile}
\alias{print.hillsfile}
\title{Print hillsfile}
\usage{
\method{print}{hillsfile}(x, ...)
}
\arguments{
\item{x}{hillsfile object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`print.hillsfile` prints dimensionality and size of a hillsfile object.
}
\examples{
acealanme
}
