% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{print.minima}
\alias{print.minima}
\title{Print minima object}
\usage{
\method{print}{minima}(x, ...)
}
\arguments{
\item{x}{minima object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`print.minima` prints free energy minima (identifier, values of bins and collective variables and free energy).
}
\examples{
tfes<-fes(acealanme, imax=5000)
minima<-fesminima(tfes)
minima
}
