% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebs.R
\name{print.nebpath}
\alias{print.nebpath}
\title{Print Nudged Elastic Band minima}
\usage{
\method{print}{nebpath}(nebpath = nebpath, ...)
}
\arguments{
\item{nebpath}{nebpath object}
}
\description{
`print.nebpath` prints the list minima for Nudged Elastic Band
}
\examples{
tfes<-fes(acealanme)
minima<-fesminima(tfes)
nebAD<-neb(minima, min1="A", min2="D")
nebAD
}
