% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{print.profiles}
\alias{print.profiles}
\title{Print profiles object}
\usage{
\method{print}{profiles}(x, ...)
}
\arguments{
\item{x}{minima object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`print.profiles` prints free energy profile.
}
\examples{
tfes<-fes(acealanme)
minima<-fesminima(tfes)
prof<-feprof(minima)
prof
}
