% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{prob.fes}
\alias{prob.fes}
\title{Calculate probability of free energy surface}
\usage{
\method{prob.fes}(object, ...)
}
\arguments{
\item{object}{fes object.}
  
\item{temp}{temperature in Kelvins}

\item{eunit}{energy units (kJ/mol or kcal/mol, kJ/mol is default)}
}
\description{
`prob.fes` calculates probability from free energy in a fes object.
}
\examples{
tfes<-fes(acealanme, imax=5000)
print(prob(tfes))
}
