% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{read.fes}
\alias{read.fes}
\title{Read free energy surface from the output of MetadynView}
\usage{
read.fes(filename = "fes.txt", dimension = 2, per = c(TRUE, TRUE),
  pcv1 = c(-pi, pi), pcv2 = c(-pi, pi))
}
\arguments{
\item{dimension}{number of dimensions (default 2)}

\item{per}{logical vector specifying periodicity of collective variables}

\item{pcv1}{periodicity of CV1}

\item{pcv2}{periodicity of CV2}

\item{file}{fes.txt file from MetadynView}
}
\value{
hillsfile object
}
\description{
`read.fes` read free energy surface from the output file of MetadynView (http://metadyn.vscht.cz)
}
\examples{
tfes<-read.fes("fes.txt")
}
