% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{summary.fes}
\alias{summary.fes}
\title{Print summary of free energy surface}
\usage{
\method{summary}{fes}(inputfes = inputfes)
}
\arguments{
\item{inputfes}{fes object}
}
\description{
`summary.fes` prints dimensionality, minimum and maximum of
free energy in a fes object
}
\examples{
tfes<-fes(acealanme)
summary(tfes)
}
