% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{summary.hillsfile}
\alias{summary.hillsfile}
\title{Print summary for hillsfile}
\usage{
\method{summary}{hillsfile}(hills = hills)
}
\arguments{
\item{hills}{hillsfile object}
}
\value{
dimensionality and size of a hillsfile object
}
\description{
`summary.hillsfile` prints dimensionality, size and collective variable ranges of a hillsfile object
}
\examples{
acealanme
}
