% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{summary.profiles}
\alias{summary.profiles}
\title{Print summary for free energy profile}
\usage{
\method{summary}{profiles}(object, imind = 1, imaxd = NULL, ...)
}
\arguments{
\item{object}{profiles object.}

\item{imind}{index of a hill from which calculation of difference
starts (default 0).}

\item{imaxd}{index of a hill from which calculation of difference
stops (default the rest of hills).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`summary.profiles` prints the list of free energy minima with maximal
and minimal free energy differences.
}
\examples{
tfes<-fes(acealanme)
minima<-fesminima(tfes)
prof<-feprof(minima)
summary(prof)
}
