% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{tail.hillsfile}
\alias{tail.hillsfile}
\title{Print last n lines of hillsfile}
\usage{
\method{tail}{hillsfile}(x, n = 10, ...)
}
\arguments{
\item{x}{hillsfile object.}

\item{n}{number of lines (default 10).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`tail.hillsfile` prints last n lines of a hillsfile object.
}
\examples{
tail(acealanme)
}
