% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acealanme1d}
\alias{acealanme1d}
\title{Hills from 30 ns metadynamics of AceAlaNme in water with one collective variable}
\format{hillsfile object}
\usage{
acealanme1d
}
\description{
Hills from 30 ns metadynamics of AceAlaNme (Amber99SB-ILDN) in water (TIP3P)
with Ramachandran dihedrals phi as collective variables.
}
\keyword{datasets}
