% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{feprof}
\alias{feprof}
\title{Calculate free energy profile for minima object (generic function for 'metadynminer'
and 'metadynminer3d')}
\usage{
feprof(minims, imax)
}
\arguments{
\item{minims}{minima object.}

\item{imax}{index of a hill from which summation stops (default the rest of hills).}
}
\description{
`feprof` calculates free energy profiles for free energy minima. It finds the global minimum
at the `imax` and calculates the evolution of free energies of a local vs. the global free energy
minimum. The free energy of the global minimum is constant (zero).
}
