% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{feprof}
\alias{feprof}
\title{Calculate free energy profile for minima object}
\usage{
feprof(minims = minims, imin = 0, imax = NULL)
}
\arguments{
\item{minims}{minima object}

\item{imin}{index of a hill from which summation starts (default 0)}

\item{imax}{index of a hill from which summation stops (default the rest of hills)}
}
\description{
`feprof` calculates free energy profiles for free energy minima. It finds the global minimum at the
imax and calculates the evolution of free of local vs. global free energy minimum. The global
minimum is const (zero).
}
\examples{
tfes<-fes(acealanme)
minima<-findminima(tfes)
prof<-fesprof(minima)
prof
}
