% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{head.hillsfile}
\alias{head.hillsfile}
\title{Print first n lines of hillsfile}
\usage{
\method{head}{hillsfile}(hills = hills, n = 10, ...)
}
\arguments{
\item{hills}{hillsfile object}

\item{n}{number of lines (default 10)}
}
\value{
first n lines of a hillsfile object
}
\description{
`head.hillsfile` prints first n lines of a hillsfile object.
}
\examples{
head(acealanme)
}
