% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{lines.fes}
\alias{lines.fes}
\title{Plots 1D fes object as lines}
\usage{
\method{lines}{fes}(inputfes = inputfes, x = NULL, lwd = 1,
  col = "black")
}
\arguments{
\item{inputfes}{fes object}
}
\description{
`lines.fes` plots 1D free energy surface as lines.
}
\examples{
tfes<-fes(acealanme1d)
plot(tfes)
lines(tfes, lwd=4)
}
