% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{min.fes}
\alias{min.fes}
\title{Calculate minimum of free energy surface}
\usage{
\method{min}{fes}(inputfes = inputfes, na.rm = NULL)
}
\arguments{
\item{inputfes}{fes object}
}
\value{
minimum
}
\description{
`min.fes` calculates minimum of free energy in a fes object
}
\examples{
tfes<-fes(acealanme)
min(tfes)
}
