% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{plot.hillsfile}
\alias{plot.hillsfile}
\title{Plot hillsfile object}
\usage{
\method{plot}{hillsfile}(hills = hills, ignoretime = FALSE,
  xlab = NULL, ylab = NULL, xlim = NULL, ylim = NULL,
  main = NULL, sub = NULL, pch = 1, col = "black", bg = "red",
  cex = 1, asp = NULL, lwd = 1, axes = TRUE)
}
\arguments{
\item{hills}{hillsfile object}

\item{ignoretime}{time in the first column of the HILLS file will be ignored}
}
\description{
`plot.hillsfile` plots hillsfile object. For a hillsfile with one collective variable it plots its evolution.
For a hillsfile with two collective variables it plots CV1 vs. CV2
}
\examples{
plot(acealanme)
}
