\name{plot.hillsfile}
\alias{plot.hillsfile}
\title{
  Plots the profile of CVs of metadynamics
}
\description{
  Plots the profile of CVs of metadynamics. For metadynamics with one CV it plots CV as a function of time.
  For metadynamics with two CVs it plots CV2 as a function of CV1.
}
\usage{
plot.hillsfile(hills)
}
\arguments{
  \item{hills}{
    hillsfile is a data frame containing time, CV(s), hill width(s), height and bias factor.
  }
}
\details{
   Other plot parameters such as \code{xlab}, \code{ylab}, \code{xlim}, \code{ylim}, \code{zlim}, \code{main}, \code{sub}, \code{pch}, \code{col}, \code{bg}, \code{cex}, \code{asp}, \code{lwd} or \code{axes} can be used.
}
\value{
}
\references{
}
\author{
Vojtech Spiwok
}
\note{
}

\seealso{
}
\examples{
  
}

