% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebs.R
\name{plot.nebpath}
\alias{plot.nebpath}
\title{Plot Nudged Elastic Band}
\usage{
\method{plot}{nebpath}(nebpath = nebpath, xlim = NULL, ylim = NULL,
  main = NULL, sub = NULL, xlab = "bin", ylab = "free energy",
  col = "red", lwd = 1, cex = 1, axes = T)
}
\arguments{
\item{nebpath}{nebpath object}
}
\description{
`plot.nebpath` plots free energy profile calculated by Nudged Elastic Band.
}
\examples{
tfes<-fes(acealanme)
minima<-findminima(tfes)
nebAD<-neb(minima, min1="A", min2="D")
plot(nebAD)
}
