% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimaandpaths.R
\name{plot.profiles}
\alias{plot.profiles}
\title{Plot free energy profile}
\usage{
\method{plot}{profiles}(profs = profs, which = NULL,
  ignoretime = FALSE, xlim = NULL, ylim = NULL, main = NULL,
  sub = NULL, xlab = NULL, ylab = NULL, col = NULL, asp = NULL,
  lwd = 1, axes = T)
}
\arguments{
\item{profs}{profiles object}

\item{which}{vector of indexes of profiles to be ploted (default all)}

\item{ignoretime}{time in the first column of the HILLS file will be ignored}
}
\description{
`plot.profiles` plots evolution of free energy differences between minima.
They are colored by rainbow colors from the global one (blue) to the highest (red).
}
\examples{
tfes<-fes(acealanme)
minima<-fesminima(tfes)
prof<-feprof(minima)
plot(prof)
}
