% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{points.hillsfile}
\alias{points.hillsfile}
\title{Plot points for hillsfile object}
\usage{
\method{points}{hillsfile}(hills = hills, ignoretime = FALSE,
  pch = 1, col = "black", bg = "red", cex = 1, asp = NULL,
  lwd = 1, axes = TRUE, ...)
}
\arguments{
\item{hills}{hillsfile object}

\item{ignoretime}{time in the first column of the HILLS file will be ignored}
}
\description{
`points.hillsfile` plots points for hillsfile object. For a hillsfile with one
collective variable it plots its evolution. For a hillsfile with two collective
variables it plots CV1 vs. CV2
}
\examples{
plot(acealanme)
points(acealanme, col="red")
}
