% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebs.R
\name{points.nebpath}
\alias{points.nebpath}
\title{Plot points for Nudged Elastic Band}
\usage{
\method{points}{nebpath}(nebpath = nebpath, pch = NULL, cex = 1,
  bg = NULL, col = "red", lwd = 1)
}
\arguments{
\item{nebpath}{nebpath object}
}
\description{
`points.nebpath` plots points for free energy profile calculated by Nudged Elastic Band.
}
\examples{
tfes<-fes(acealanme)
minima<-findminima(tfes)
nebAD<-neb(minima, min1="A", min2="D")
plot(nebAD)
points(nebAD)
}
