% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{print.fes}
\alias{print.fes}
\title{Print dimensionality, minimum and maximum of free energy surface}
\usage{
\method{print}{fes}(x, ...)
}
\arguments{
\item{x}{fes object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`print.fes` prints dimensionality, minimum and maximum of
free energy in a fes object
}
\examples{
tfes<-fes(acealanme)
tfes
}
