% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{print.hillsfile}
\alias{print.hillsfile}
\title{Print hillsfile}
\usage{
\method{print}{hillsfile}(hills = hills)
}
\arguments{
\item{hills}{hillsfile object}
}
\value{
dimensionality and size of a hillsfile object
}
\description{
`print.hillsfile` prints dimensionality and size of a hillsfile object.
}
\examples{
acealanme
}
