% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{print.minima}
\alias{print.minima}
\title{Print minima object}
\usage{
\method{print}{minima}(minims)
}
\arguments{
\item{minims}{minima object}

\item{temp}{temperature in Kelvins}

\item{eunit}{energy units (kJ/mol or kcal/mol, kJ/mol is default)}
}
\description{
`print.minima` prints free energy minima (identifier, values of bins and collective variables and free energy).
}
\examples{
tfes<-fes(acealanme)
minima<-findminima(tfes)
minima
}
