% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebs.R
\name{print.nebpath}
\alias{print.nebpath}
\title{Print Nudged Elastic Band minima}
\usage{
\method{print}{nebpath}(x, ...)
}
\arguments{
\item{x}{nebpath object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`print.nebpath` prints the list minima for Nudged Elastic Band
}
\examples{
tfes<-fes(acealanme, imax=5000)
minima<-fesminima(tfes)
nebAD<-neb(minima, min1="A", min2="D", nsteps=20)
nebAD
}
