% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{prob}
\alias{prob}
\title{Calculate probability of free energy surface}
\usage{
prob(inputfes, temp = 300, eunit = "kJ/mol")
}
\arguments{
\item{inputfes}{fes object.}

\item{temp}{temperature in Kelvins (default 300).}

\item{eunit}{energy units (kJ/mol or kcal/mol, kJ/mol is default).}
}
\description{
`prob` calculates probability from free energy in a fes object.
}
\examples{
tfes<-fes(acealanme, imax=5000)
print(prob(tfes))
}
