% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{read.hills}
\alias{read.hills}
\title{Read HILLS from Plumed}
\usage{
read.hills(file = "HILLS", per = c(FALSE, FALSE), pcv1 = c(-pi, pi),
  pcv2 = c(-pi, pi), ignoretime = FALSE)
}
\arguments{
\item{file}{HILLS file from Plumed}

\item{per}{logical vector specifying periodicity of collective variables}

\item{pcv1}{periodicity of CV1}

\item{pcv2}{periodicity of CV2}

\item{ignoretime}{time in the first column of the HILLS file will be ignored}
}
\value{
hillsfile object
}
\description{
`read.hills` reads a HILLS file generated by Plumed and returns a hillsfile object.
User can specify whether some collective variables are periodic.
}
\examples{
hillsf <- read.hills("HILLS", per=c(T, T))
}
