% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{summary.hillsfile}
\alias{summary.hillsfile}
\title{Print summary for hillsfile}
\usage{
\method{summary}{hillsfile}(object, ...)
}
\arguments{
\item{object}{hillsfile object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
`summary.hillsfile` prints dimensionality, size and collective variable ranges of a hillsfile object.
}
\examples{
summary(acealanme)
}
