% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{summary.minima}
\alias{summary.minima}
\title{Print minima object summary}
\usage{
\method{summary}{minima}(minims = minims, temp = 300,
  eunit = "kJ/mol")
}
\arguments{
\item{minims}{minima object}

\item{temp}{temperature in Kelvins}

\item{eunit}{energy units (kJ/mol or kcal/mol, kJ/mol is default)}
}
\description{
`summary.minima` prints summary for free energy minima (identifier, values of bins and collective variables,
free energy and equilibrium populations).
}
\examples{
tfes<-fes(acealanme)
minima<-findminima(tfes)
summary(minima)
}
