% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{summary.nebpath}
\alias{summary.nebpath}
\title{Print summary for Nudged Elastic Band}
\usage{
\method{summary}{nebpath}(nebpath = nebpath, temp = 300,
  eunit = "kJ/mol")
}
\arguments{
\item{nebpath}{nebpath object}
}
\description{
`print.nebpath` prints the list minima for Nudged Elastic Band, activation energies and
half lives calculated by Eyring equation (https://doi.org/10.1063/1.1749604).
}
\examples{
tfes<-fes(acealanme)
minima<-findminima(tfes)
nebAD<-neb(minima, min1="A", min2="D")
summary(nebAD)
}
