% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadynminer.R
\name{summary.profiles}
\alias{summary.profiles}
\title{Print summary for free energy profile}
\usage{
\method{summary}{profiles}(profs = profs, imind = 1, imaxd = NULL)
}
\arguments{
\item{profs}{profiles object}

\item{imind}{index of a hill from which calculation of difference starts (default 0)}

\item{imaxd}{index of a hill from which calculation of difference stops (default the rest of hills)}
}
\description{
`summary.profiles` prints the list of free energy minima with maximal
and minimal free energy differences.
}
\examples{
tfes<-fes(acealanme)
minima<-findminima(tfes)
prof<-fesprof(minima)
summary(prof)
}
