% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readingandfes.R
\name{tail.hillsfile}
\alias{tail.hillsfile}
\title{Print last n lines of hillsfile}
\usage{
\method{tail}{hillsfile}(hills = hills, n = 10, ...)
}
\arguments{
\item{hills}{hillsfile object}

\item{n}{number of lines (default 10)}
}
\value{
last n lines of a hillsfile object
}
\description{
`tail.hillsfile` prints last n lines of a hillsfile object.
}
\examples{
tail(acealanme)
}
